src/main/java/com/example/blog/service/impl/UserServiceImpl.java
package com.example.blog.service.impl;

import com.example.blog.models.Role;
import com.example.blog.models.User;
import com.example.blog.repository.RoleRepository;
import com.example.blog.repository.UserRepository;
import com.example.blog.service.UserService;
import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
import org.springframework.stereotype.Service;

@Service
public class UserServiceImpl implements UserService {
    private final UserRepository userRepository;
    private final RoleRepository roleRepository;
    private final BCryptPasswordEncoder passwordEncoder;

    public UserServiceImpl(BCryptPasswordEncoder bCryptPasswordEncoder, RoleRepository roleRepository, UserRepository userRepository) {
        this.passwordEncoder = bCryptPasswordEncoder;
        this.roleRepository = roleRepository;
        this.userRepository = userRepository;
    }

    public User save(User user) {
        Long roleId = user.getRole().getId();
        Role userRole = roleRepository.findById(roleId)
                .orElseThrow(() -> new RuntimeException(String.format("Role (id = %d) not found", roleId)));
        user.setRole(userRole);
        user.setPassword(passwordEncoder.encode(user.getPassword()));
        return userRepository.save(user);
    }

    public User findByUserName(String login) {
        return userRepository.findByuserName(login);
    }

    public User findByUserNameAndPassword(String login, String password) {
        User userEntity = findByUserName(login);
        if (userEntity != null) {
            if (passwordEncoder.matches(password, userEntity.getPassword())) {
                return userEntity;
            }
        }
        return null;
    }
}
